# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gerenciamento de fazendas leiteiras\nSistema completo para gestão de animais, coletas de leite, reproduções, vendas e dívidas",
        "title": "FazendaPro API",
        "contact": {
            "name": "Suporte FazendaPro",
            "email": "suporte@fazendapro.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/animals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados de um animal específico pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Obter animal por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do animal",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AnimalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de um animal existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Atualizar animal",
                "parameters": [
                    {
                        "description": "Dados atualizados do animal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAnimalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AnimalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria um novo animal na fazenda",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Criar animal",
                "parameters": [
                    {
                        "description": "Dados do animal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAnimalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um animal do sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Deletar animal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do animal",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/animals/farm": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna lista de todos os animais de uma fazenda específica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Obter animais por fazenda",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da fazenda",
                        "name": "farmId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.AnimalResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/animals/photo": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Faz upload de uma foto para um animal específico",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Upload foto do animal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do animal",
                        "name": "animal_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Arquivo de imagem",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AnimalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/animals/sex": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna lista de animais de uma fazenda filtrados por sexo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Obter animais por sexo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da fazenda",
                        "name": "farmId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Sexo (1=Fêmea, 2=Macho)",
                        "name": "sex",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.AnimalResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateAnimalRequest": {
            "description": "Dados necessários para criar um novo animal",
            "type": "object",
            "properties": {
                "animal_name": {
                    "description": "Nome do animal",
                    "type": "string",
                    "example": "Branquinha"
                },
                "animal_type": {
                    "description": "Tipo de animal (enum)",
                    "type": "integer",
                    "example": 1
                },
                "birth_date": {
                    "description": "Data de nascimento (ISO format)",
                    "type": "string",
                    "example": "2020-01-15"
                },
                "breed": {
                    "description": "Raça",
                    "type": "string",
                    "example": "Holandesa"
                },
                "castrated": {
                    "description": "Se está castrado",
                    "type": "boolean",
                    "example": false
                },
                "confinement": {
                    "description": "Se está em confinamento",
                    "type": "boolean",
                    "example": false
                },
                "current_batch": {
                    "description": "Lote atual",
                    "type": "integer",
                    "example": 1
                },
                "ear_tag_number_local": {
                    "description": "Número da brinco local",
                    "type": "integer",
                    "example": 123
                },
                "ear_tag_number_register": {
                    "description": "Número de registro",
                    "type": "integer",
                    "example": 456
                },
                "farm_id": {
                    "description": "ID da fazenda",
                    "type": "integer",
                    "example": 1
                },
                "father_id": {
                    "description": "ID do pai (opcional)",
                    "type": "integer",
                    "example": 10
                },
                "fertilization": {
                    "description": "Se está fertilizada",
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "description": "ID único do animal",
                    "type": "integer",
                    "example": 1
                },
                "mother_id": {
                    "description": "ID da mãe (opcional)",
                    "type": "integer",
                    "example": 20
                },
                "photo": {
                    "description": "Foto em base64",
                    "type": "string",
                    "example": "data:image/jpeg;base64,..."
                },
                "purpose": {
                    "description": "Propósito do animal (enum)",
                    "type": "integer",
                    "example": 1
                },
                "sex": {
                    "description": "Sexo (1=Fêmea, 2=Macho)",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "description": "Status do animal (enum)",
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "description": "Tipo de animal",
                    "type": "string",
                    "example": "Bovino"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e depois o token JWT. Exemplo: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}